dz-4-1

+++ src/controllers/notesController.js
@@ -8,7 +8,7 @@ export const getAllNotes = async (req, res) => {
 
   const skip = (page - 1) * perPage;
 
-  const notesQuery = Note.find();
+  const notesQuery = Note.find({userId: req.user._id});
 
   if (tag) {
     notesQuery.where("tag").equals(tag);
@@ -37,7 +37,7 @@ export const getAllNotes = async (req, res) => {
 export const getNoteById = async (req, res, next) => {
   const { noteId } = req.params;
 
-  const note = await Note.findById(noteId);
+  const note = await Note.findById({noteId, userId: req.user._id});
 
   if (!note) {
     next(createHttpError(404, "Note not found"));
@@ -49,7 +49,7 @@ export const getNoteById = async (req, res, next) => {
 
 export const createNote = async (req, res) => {
 
-  const newNote = await Note.create(req.body);
+  const newNote = await Note.create({ ...req.body, userId: req.user._id });
 
   res.status(201).json(newNote);
 };
@@ -57,7 +57,7 @@ export const createNote = async (req, res) => {
 export const deleteNote = async (req, res, next) => {
   const { noteId } = req.params;
 
-  const note = await Note.findOneAndDelete({ _id: noteId });
+  const note = await Note.findOneAndDelete({ _id: noteId, userId:  req.user._id });
 
   if (!note) {
     next(createHttpError(404, "Note not found"));
@@ -71,7 +71,7 @@ export const updateNote = async (req, res, next) => {
   const { noteId } = req.params;
 
   const note = await Note.findOneAndUpdate(
-    { _id: noteId },
+    { _id: noteId, userId:  req.user._id },
     req.body,
     { new: true }
   );
